microservicios-ia/descriptions/requirements.txt
Flask
openai
requests
python-dotenv

